% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.r
\name{assignment}
\alias{assignment}
\title{Custom RMarkdown document type that parses student answers to assignments}
\usage{
assignment(pkg, ...)
}
\arguments{
\item{pkg}{path to course package for this assignment}

\item{...}{document parameters}
}
\value{
an R Markdown output format definition
}
\description{
Flag an Rmd file as containing a set of answers to an assignment by setting
its output type to this function; not meant to be called directly.
}
