% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseR.r
\name{build}
\alias{build}
\title{Build a project}
\usage{
build(cleanBuild = FALSE, cleanPreviews = TRUE, path = getwd())
}
\arguments{
\item{cleanBuild}{should intermediate build files be deleted when the build 
is done.  Note this will trounce any knitr caching}

\item{cleanPreviews}{should .html files created in the main project directory
by making ad hoc previews in RStudio be deleted}

\item{path}{path for the project to build}
}
\description{
Build a project
}

