% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.r
\name{check_final}
\alias{check_final}
\title{Short hand for checkr test final_() followed by check_value().}
\usage{
check_final(...)
}
\arguments{
\item{...}{argunents passed along to \code{\link{checkr::check_value}}}
}
\value{
test function that takes a capture
}
\description{
Short hand for checkr test final_() followed by check_value().
}
