% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{check_rows}
\alias{check_rows}
\title{Code checking function that ensures the last value in a set of expressions is
a table with the same number of rows as a target}
\usage{
check_rows(x)
}
\arguments{
\item{x}{the target table}
}
\value{
test function that takes a capture
}
\description{
Code checking function that ensures the last value in a set of expressions is
a table with the same number of rows as a target
}
