% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{checkAssignment}
\alias{checkAssignment}
\title{Check an assignment}
\usage{
checkAssignment(name, path = getwd(), autoknit = TRUE, pkg)
}
\arguments{
\item{name}{name of the assignment to check}

\item{path}{optional path to your course project folder}

\item{autoknit}{if TRUE your assignment Rmd files will automatically be
re-knitted when you save changes}

\item{pkg}{path to course package}
}
\value{
true if we don't find anything dodgy
}
\description{
Allows you to check your current progress on an assignment against a 
reference solution.
}
