% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseR.r
\name{clean}
\alias{clean}
\title{Delete intermediate and final build files from the project}
\usage{
clean(path = getwd())
}
\arguments{
\item{path}{project path}
}
\description{
Delete intermediate and final build files from the project
}
