% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{copyErrorLogs}
\alias{copyErrorLogs}
\title{Copies any current error logs for your published shiny app into your current
working directory. Note the logs are only available while your app is still
running; once you close the window they will be erased.}
\usage{
copyErrorLogs(projectName, pkg)
}
\arguments{
\item{projectName}{string with your group's project name.  This must match 
your group's repository name on GitHub.}

\item{pkg}{path to course package}
}
\description{
Copies any current error logs for your published shiny app into your current
working directory. Note the logs are only available while your app is still
running; once you close the window they will be erased.
}
