\name{courseR.init}
\alias{courseR.init}
\title{Initialize a new courseR project directory}
\usage{
courseR.init(path = getwd(), overwrite = FALSE)
}
\arguments{
  \item{path}{character string path on the local file
  system where a new project folder should be written.
  Defaults to the current working directory.}

  \item{overwrite}{a logical value specifying whether or
  not existing files should be overwritten.}
}
\description{
After initializing a new project see the 'courseR.json'
file and '/content' subdirectory for examples.
}

