% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.r
\name{createTable}
\alias{createTable}
\title{Create a table source file}
\usage{
createTable(nrows, cols, row.col = "#", file)
}
\arguments{
\item{nrows}{number of rows}

\item{cols}{column names}

\item{row.col}{name of column for row numbers; set to false to skip}

\item{file}{output YAML file}
}
\value{
the YAML string invisibly
}
\description{
The resulting YAML file can be easily hand editted and included as a table in
an Rmd file with the code{\link{renderTable}} function.
}
