% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.r
\name{includeTable}
\alias{includeTable}
\title{Render a table in an Rmd chunk}
\usage{
includeTable(file, markdown = TRUE, template = "templates/site/table.html",
  widths = NULL)
}
\arguments{
\item{file}{source YAML file}

\item{markdown}{if TRUE text values in the YAML file are preprocessed as 
markdown}

\item{template}{table template file}

\item{widths}{if not null gives the grid system widths of the columns}
}
\value{
the HTML table tagged by htmltools so that it's rendered as HTML when
  an the rmd is knitted.
}
\description{
Use this function with the YML template file created by 
code{\link{createTable}}.
}
