% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/courseR.R
\name{manifest}
\alias{manifest}
\title{Get project content manifest}
\usage{
manifest(projectPath = getwd(), config = yaml.load_file(configPath),
  configPath = file.path(projectPath, "courseR.yaml"),
  buildCache = file.path(projectPath, config$paths$buildCache))
}
\arguments{
\item{projectPath}{a character string with the root path to a project folder.
This folder should contain the required resource subfolders (unless
alternative locations are specified). Defaults to the current working
directory.}

\item{config}{a list structure containing a full set of project configuration
settings.  Defaults to the yaml parse of the file found on configPath.}

\item{configPath}{a character string path to a yaml configuration file.
Defaults to courseR.yaml.}

\item{buildCache}{a character string containing the path to the build cache
  directory.}
}
\value{
A data.frame
}
\description{
Get project content manifest
}

