% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseR.r
\name{newAssignment}
\alias{newAssignment}
\title{Create a new assignment file}
\usage{
newAssignment(file, title, data = "", path = getwd(), ...)
}
\arguments{
\item{file}{the name of the new file to create}

\item{title}{the title of the new assignment}

\item{data}{files in the data/ folder that this assignment relies upon}

\item{path}{path to the project to create a new assignment in}

\item{...}{additional arguments are available as keys when the template is
rendered to create the new file}
}
\description{
Create a new assignment file
}
