% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseR.r
\name{newContent}
\alias{newContent}
\title{Create a new course content file}
\usage{
newContent(file, title, data = "", path = getwd(), ...)
}
\arguments{
\item{file}{the name of the new file to create}

\item{title}{the title of the new piece of course content}

\item{data}{files in the data/ folder that this content relies upon}

\item{path}{path to the project to create a new content file in}

\item{...}{additional arguments are available as keys when the template is
rendered to create the new file}
}
\description{
Create a new course content file
}
