% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{publishApp}
\alias{publishApp}
\title{Publish the current version of a Shiny app project hosted on the course's 
Bio-185 GitHub organization to the RNA Shiny server.}
\usage{
publishApp(projectName, remove = TRUE, pkg)
}
\arguments{
\item{projectName}{string with your group's project name.  This must match
your group's repository name on GitHub.}

\item{remove}{if TRUE will remove the old version of your app and update.}

\item{pkg}{path to course package}
}
\description{
Publish the current version of a Shiny app project hosted on the course's 
Bio-185 GitHub organization to the RNA Shiny server.
}
