% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.r
\name{renderTable}
\alias{renderTable}
\title{Render a table in an Rmd chunk}
\usage{
renderTable(file, markdown = TRUE, template = "templates/site/table.html",
  widths = NULL)
}
\arguments{
\item{file}{source YAML file}

\item{markdown}{if TRUE text values in the YAML file are preprocessed as
markdown}

\item{template}{table template file}

\item{widths}{if not null gives the grid system widths of the columns}
}
\value{
the HTML table taged as HTML tools tags so that it's rendered as HTML
  with the rmd is knit
}
\description{
Use this function with the file created by code{\link{createTable}}
}
