% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.r
\name{solution}
\alias{solution}
\title{Custom RMarkdown document type that parses assignment solution files}
\usage{
solution(...)
}
\arguments{
\item{...}{document parameters}
}
\value{
an R Markdown output format definition
}
\description{
Flag an Rmd file as solutions by setting its output type to this function;
not meant to be called directly.
}
