% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{startAssignment}
\alias{startAssignment}
\title{Start an assignment}
\usage{
startAssignment(name, overwrite = FALSE, path = getwd(), pkg)
}
\arguments{
\item{name}{name of the assignment to start}

\item{overwrite}{if TRUE this function will overwrite any existing work 
you've done on this asignment already}

\item{path}{optional path to your course project folder}

\item{pkg}{path to course package}
}
\value{
new file path
}
\description{
This function will create an RMD file that you can use to start working on
solutions for an assignment.  It also refreshes the data files in your data/
directory.
}

