% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.r
\name{taskCollector}
\alias{taskCollector}
\title{Collect task chunks/source code blocks}
\usage{
taskCollector(callback, rds)
}
\arguments{
\item{callback}{tail call callback}

\item{rds}{target data file}
}
\value{
function to set as knitr hook
}
\description{
knitr hook function to collect source and result chunks when assignments are
knit
}

