% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseR.r
\name{update}
\alias{update}
\title{Update a project}
\usage{
update(path = getwd())
}
\arguments{
\item{path}{the path for the project to update}
}
\description{
Updates files that depend on the current contents.  This is called
automatically from \code{\link{newContent}} and \code{\link{newAssignment}}.
}

