% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{website}
\alias{website}
\title{Open course website}
\usage{
website(pkg)
}
\arguments{
\item{pkg}{path to course package}
}
\description{
This convenience function will open the course website in the RStudio Viewer
panel.  Only works if run from within RStudio.
}
