% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.r
\name{assignment}
\alias{assignment}
\title{Custom RMarkdown document type that parses student assignments}
\usage{
assignment(assignment, submit = FALSE, ...)
}
\arguments{
\item{assignment}{name of the assignment}

\item{submit}{bool is this a submission render?}

\item{...}{}
}
\value{
an R Markdown output format definition
}
\description{
Flag an Rmd file as an assignment by setting its output type to this
function; not meant to be called directly.
}

