% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseR.r
\name{build}
\alias{build}
\title{Build a project}
\usage{
build(path = getwd(), cleanBuild = FALSE, cleanPreviews = TRUE)
}
\arguments{
\item{path}{path for the project to build.}

\item{cleanBuild}{should intermediate build files be deleted when the build 
is done.  Note this will trounce any knitr caching.}

\item{cleanPreviews}{should .html files created in the main project directory
by making ad hoc previews in RStudio be deleted.}
}
\description{
Build a project
}
