% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseR.r
\name{build}
\alias{build}
\title{Build a project}
\usage{
build(path = getwd(), cleanBuild = FALSE, cleanPreviews = TRUE)
}
\arguments{
\item{path}{path for the project to build.}

\item{cleanBuild}{if TRUE completely rebuilds the project; erases knitr 
caches and the contents of 'build/' and 'dist/'.  Set to TRUE if you have
changed any of the website templates since the last build.}

\item{cleanPreviews}{should .html files created in the main project directory
by making ad hoc previews in RStudio be deleted.}
}
\description{
Build a project
}
