% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{checkAssignment}
\alias{checkAssignment}
\title{Check an assignment}
\usage{
checkAssignment(name, path = getwd())
}
\arguments{
\item{name}{name of the assignment to check}

\item{path}{path to your course project folder}
}
\value{
true if we don't find anything dodgy
}
\description{
Allows you to check your current progress on an assignment against a
reference solution.
}

