% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{checkAssignments}
\alias{checkAssignments}
\title{Check assignments}
\usage{
checkAssignments(path = getwd())
}
\arguments{
\item{path}{}
}
\description{
This function behaves differently when called from a student versus 
instructor account.  For students, it shows your current progress on all
assignments; for instructors a grading interface.
}

