% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/courseR.R
\name{init}
\alias{init}
\title{Initialize a new courseR project directory}
\usage{
init(path = getwd(), overwrite = FALSE)
}
\arguments{
\item{path}{character string path on the local file system where a new
project folder should be written.  Defaults to the current working
directory.}

\item{overwrite}{a logical value specifying whether or not existing files
should be overwritten.}
}
\description{
After initializing a new project see the '/content' subdirectory for example
content pages.  See the 'courseR.yaml' configuration file and /templates'
subdirectory for ways to customize the visual look and structure of the
project site.
}
\details{
Run (\code{\link{courseR.build}}) to render and build the site.
}

