% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{listAssignments}
\alias{listAssignments}
\title{List available assignments}
\usage{
listAssignments(pkg)
}
\arguments{
\item{pkg}{path to course package}

\item{path}{optional path to your course project folder}
}
\value{
character vector with assignment names
}
\description{
List available assignments
}

