% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{startAssignment}
\alias{startAssignment}
\title{Start an assignment}
\usage{
startAssignment(name, overwrite = FALSE, path = getwd())
}
\arguments{
\item{name}{name of the assignment to start}

\item{overwrite}{if true this function will overwrite any existing work
you've done on this asignment already}

\item{path}{path to your course project folder}
}
\description{
Start an assignment
}

