% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{submitAssignment}
\alias{submitAssignment}
\title{Submit an assignment}
\usage{
submitAssignment(name, path = getwd())
}
\arguments{
\item{name}{name of the assignment to submit}

\item{path}{path to your course project folder}
}
\description{
Submitting an assignment flags the current version of your work as complete 
and/or ready for grading.  You can resubmit assignments; doing so will reset
the current grading information.
}

